requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.8.4
neo4j==5.19.0
python-dotenv==1.0.1
pydantic==2.6.4
